+++ ltrace.cpp
@@ -179,7 +179,7 @@ bool LTrace::Trace_Light(LRoomManager &manager, const LLight &light, eLightRun e
 	const LSource &cam = light.m_Source;
 
 	unsigned int pool_member = manager.m_Pool.Request();
-	assert (pool_member != -1);
+	assert (pool_member != (unsigned int) -1);
 
 	LVector<Plane> &planes = manager.m_Pool.Get(pool_member);
 	planes.clear();
@@ -515,7 +515,7 @@ void LTrace::Trace_Recursive(int depth, LRoom &room, const LVector<Plane> &plane
 
 		// else recurse into that portal
 		unsigned int uiPoolMem = LMAN->m_Pool.Request();
-		if (uiPoolMem != -1)
+		if (uiPoolMem != (unsigned int) -1)
 		{
 			// get a vector of planes from the pool
 			LVector<Plane> &new_planes = LMAN->m_Pool.Get(uiPoolMem);
